allow selecting existing upgrades in modal, and move some fucntions from view to model
