+++ js-src/modules/models/shipBuild.js
@@ -3,8 +3,8 @@
 var _ = require('lodash');
 var ships = require('./ships');
 var events = require('../controllers/events');
-var XpItem = require('./xpItem');
-var itemTypes = require('./itemTypes');
+var XpItem = require('./shipBuild/xpItem');
+var itemTypes = require('./shipBuild/itemTypes');
 var upgrades = require('../models/upgrades').all;
 var pilots = require('../models/pilots').allRebels;
 var EnemyDefeatsModel = require('../models/enemyDefeats');
