+++ js-src/modules/models/shipBuild/upgradeSlots.js
@@ -1,0 +1,102 @@
+'use strict';
+
+var _ = require('lodash');
+
+module.exports = {
+    getShipSlots: function (currentShip) {
+        // elite slots are dependent on pilot level
+
+        var usableUpgrades = _.map(currentShip.upgradeSlots, function (upgradeSlot) {
+            return {
+                type: upgradeSlot
+            };
+        });
+
+        usableUpgrades = usableUpgrades.concat([
+            {
+                type: 'Title'
+            },
+            {
+                type: 'Modification'
+            },
+            {
+                type: 'Modification',
+                pilotSkill: 4
+            },
+            {
+                type: 'Modification',
+                pilotSkill: 6
+            },
+            {
+                type: 'Modification',
+                pilotSkill: 8
+            },
+            {
+                type: 'Elite',
+                pilotSkill: 3
+            },
+            {
+                type: 'Elite',
+                pilotSkill: 5
+            },
+            {
+                type: 'Elite',
+                pilotSkill: 7
+            },
+            {
+                type: 'Elite',
+                pilotSkill: 9
+            }
+        ]);
+
+        return usableUpgrades;
+    },
+    getSlotsFromUpgrades: function (usableUpgrades, startingUpgrades, upgradesByType) {
+        var additionalSlotTypes = [];
+
+        // array to track which upgrades we've processed grants for, and prevent infinite loop
+        var processedGrantForIds = [];
+        var processGrants = function (upgrade) {
+            // Is the type of this upgrade allowed on this ship?
+            var foundGrant = false;
+            if (processedGrantForIds.indexOf(upgrade.id) === -1) {
+                // Only process this if we haven't already done so
+                var slotType = upgrade.slot;
+                if (usableUpgrades[slotType] || additionalSlotTypes.indexOf(slotType) >= 0) {
+                    // slot is allowed on ship, so lets process any additional slots the upgrade grants
+                    if (upgrade.grants) {
+                        _.each(upgrade.grants, function (grant) {
+                            if (grant.type === 'slot') {
+                                foundGrant = true;
+                                additionalSlotTypes.push(grant.name);
+                            }
+                        });
+                    }
+                }
+            }
+            return foundGrant;
+        };
+
+        var processGrantsList = function (upgradeList) {
+            var found = false;
+            // keep looping while we find results. If we find a grant, we have to process the
+            // whole list again, in case it affects another upgrade
+            do {
+                found = _.find(upgradeList, processGrants);
+                if (found && found.id) {
+                    processedGrantForIds.push(found.id);
+                }
+            } while (found);
+        };
+
+        // Do any starting upgrade grants before the purchased ones
+        processGrantsList(startingUpgrades);
+
+        // Add slots given by upgrades/starting upgrades
+        _.each(upgradesByType, function (upgradesList) {
+            processGrantsList(upgradesList);
+        });
+
+        return additionalSlotTypes;
+    }
+};
